requirements.txt
numpy>=1.23.0
pandas==1.3.3
matplotlib==3.4.3
Pillow==8.3.1
opencv-python==4.5.3.20210927
scikit-learn==0.24.2
streamlit
